requirements.txt
edge_impulse_linux
pyserial
numpy
